Download a minimal dataset of turtle observations from WAStD

Download a minimal dataset of turtle observations from WAStD

download_minimal_wastd_turtledata(
  source = "odk",
  year = NULL,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)

Arguments

source

The WAStD Encounter source, default: "odk".

year

The earliest calendar year to be returned. Default: NULL. This value, if not NULL, will be used to construct a query parameter when_year_gte=2019

api_url

(character) The WAStD API URL, default get_wastdr_api_url, see wastdr_setup

api_token

(character) The WAStD API token, default get_wastdr_api_token, see wastdr_setup

verbose

Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

A list of tibbles:

users A tibble of user names, roles, and contact details which can be used to resolve submitted user names to WAStD user IDs

enc A tibble of encounters (source, source ID, QA status) which can be used to determine which encounters to create / update / skip.

surveys A tibble of surveys reconstructed from source and year.

See also