Download all turtle data from DBCA's ODK Central

download_odkc_turtledata_2019(
  local_dir = here::here("media"),
  prod = "https://odkcentral.dbca.wa.gov.au",
  uat = "https://odkcentral-uat.dbca.wa.gov.au",
  tz = "Australia/Perth",
  download = TRUE,
  odkc_version = ruODK::get_default_odkc_version(),
  verbose = wastdr::get_wastdr_verbose()
)

Arguments

local_dir
A local directory to which to save the attachment files. Attachment filepaths will be relative to the directory. The directory and its parent folders will be created if not existing.

prod
The ODKC PROD server, default: "https://odkcentral.dbca.wa.gov.au".

uat
The ODKC UAT server, default: "https://odkcentral-uat.dbca.wa.gov.au".

tz
The lubridate timezone, default: "Australia/Perth

download
<lgl> Whether to download attachments to `local_dir` or not,

odkc_version
The numeric ODK Central version, e.g. 0.7 or 0.8.

verbose
Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

An S3 class "odkc_turtledata" with items:

downloaded_on An UTC POSIXct timestamp of the data snapshot.

tracks The turtle tracks and nests from form "Turtle Track or Nest 1.0".

tracks_dist Individual disturbances recorded against tracks, one row per disturbance.

tracks_log Individual nest tags recorded against nests, one row per tag.

tracks_egg Next excavation photos, one row per photo.

tracks_hatch Turtle hatchling morphometrics, one row per measured hatchling.

tracks_fan_outlier Individual hatchling track outliers recorded against hatched nests, one row per outlier.

tracks_light Individual light sources known at hatchling emergence, one row per light source.

track_tally A line transect tally of turtle tracks from form "Track Tally 0.6".

dist The disturbance and predation records from form "Predator or Disturbance 1.1".

mwi Strandings and rescues from the form "Marine Wildlife Incident 0.6 ".

mwi_dmg Individual injuries recorded against mwi, one record per injury.

mwi_tag Individual tags sighted during an mwi, one record per tag.

tsi Turtle Sightings from form "Turtle Sighting 0.1/0.2", one row per sighted turtle.

svs Survey start points from form "Site Visit Start 1.3".

sve Survey end points from form "Site Visit End 1.2".

sites An sf object of known WAStD sites.

areas An sf object of known WAStD localities.

See also