Download all turtle data from DBCA's ODK Central 2020

download_odkc_turtledata_2020(
  local_dir = here::here("media"),
  prod = "https://odkc.dbca.wa.gov.au",
  tz = ruODK::get_default_tz(),
  download = TRUE,
  odkc_version = ruODK::get_default_odkc_version(),
  verbose = wastdr::get_wastdr_verbose()
)

Arguments

local_dir

A local directory to which to save the attachment files. Attachment filepaths will be relative to the directory. The directory and its parent folders will be created if not existing.

prod

The ODKC PROD server, default: "https://odkc.dbca.wa.gov.au".

tz

The lubridate timezone, default: "Australia/Perth

download

(lgl) Whether to download attachments to local_dir or not,

odkc_version

The numeric ODK Central version, e.g. 0.7 or 0.8.

verbose

Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

An S3 class "odkc_turtledata" with items:

downloaded_on An UTC POSIXct timestamp of the data snapshot.

tracks The turtle tracks and nests from form "Turtle Track or Nest 1.0".

tracks_dist Individual disturbances recorded against tracks, one record per disturbance.

tracks_log Individual nest tags recorded against nests, one record per tag.

tracks_egg Next excavation photos, one record per photo.

tracks_hatch Turtle hatchling morphometrics, one record per measured hatchling.

tracks_fan_outlier Individual hatchling track outliers recorded against hatched nests, one record per outlier.

tracks_light Individual light sources known at hatchling emergence, one record per light source.

track_tally A line transect tally of turtle tracks from form "Track Tally 0.6".

dist The disturbance and predation records from form "Predator or Disturbance 1.1".

mwi Strandings and rescues from the form "Marine Wildlife Incident 0.6 ".

mwi_dmg Individual injuries recorded against mwi, one record per injury.

mwi_tag Individual tags sighted during an mwi, one record per tag.

tsi Turtle Sightings from form Turtle-Sighting, one record per sighted turtle.

tt Individual turtles encountered and tagged after nesting during night time turtle tagging from form Turtle-Tagging-3-0, one record per encountered turtle.

tt_dmg Characteristic damages or injuries recorded during turtle tagging, one record per damage.

tt_tag Individual tags recorded during turtle tagging, one record per tag. Tags can be of different type, re-sighted, applied, or removed.

tt_log Individual loggers recorded during turtle tagging, one record per logger. Loggers can be of different type, re-sighted, deployed, or removed.

tt_fix Turtle Tagging records with missing coordinates. These need to be backfilled into WAStD by hand from paper datasheets.

svs Survey start points from form Site-Visit-Start.

sve Survey end points from form Site-Visit-End.

sites An sf object of known WAStD sites.

areas An sf object of known WAStD localities.

See also