List sites with more than one production survey on a given date

Sites are expected to be surveyed not more than once daily with the exception of turtle tagging, where a morning survey and night tagging can happen legitimately on the same calendar date.

duplicate_surveys(surveys)

Arguments

surveys

(tibble) The output of parse_surveys.

Details

QA operators will want to open each link from this list of potential duplicate surveys, inspect their details, then decide on whether to make one the main survey and close others as duplicates with the "make production" button.

See also