Filter odkc_data to a given area_name.

Filter odkc_data to a given area_name.

filter_odkc_turtledata(
  data,
  area_name = NULL,
  verbose = wastdr::get_wastdr_verbose()
)

Arguments

data
<list> The output of data("odkc_data"), a list of tibbles and sf

area_name
<chr> The name of the area to filter the data by. Options: * NULL (default): do not filter data, return unfiltered data. * "All turtle programs": do not filter data, return unfiltered data. * "Other": Filter data to area_name NA. * Any name in `unique(odkc_data$areas$area_name)`: return data filtered to this area_name.

verbose
Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

<list> The input data, optionally filtered to a subset of records.

Examples

data(odkc_data)

# All data
data_all <- filter_odkc_turtledata(odkc_data)
nrow(data_all$tracks)

#> [1] 6

# One area
area_names <- unique(odkc_data$areas$area_name)
area_names

#>  [1] "Cape Dommett"                                    
#>  [2] "Smokey Bay Area"                                 
#>  [3] "Cable Beach Broome"                              
#>  [4] "Port Hedland"                                    
#>  [5] "Delambre Island"                                 
#>  [6] "Rosemary Island"                                 
#>  [7] "Conzinc Bay"                                     
#>  [8] "West Pilbara Turtle Program beaches Wickam"      
#>  [9] "West Pilbara Turtle Program beaches Caravan Park"
#> [10] "Barrow Island"                                   
#> [11] "Thevenard Island"                                
#> [12] "Ningaloo"                                        
#> [13] "Onslow"                                          
#> [14] "Perth Metro"                                     

data_area_1 <- filter_odkc_turtledata(odkc_data, area_name = "Cape Dommett")
nrow(data_area_1$tracks)

#> [1] 0

data_area_2 <- filter_odkc_turtledata(odkc_data, area_name = area_names[[2]])
nrow(data_area_2$tracks)

#> [1] 0

# Orphans
data_orphans <- filter_odkc_turtledata(odkc_data, area_name = "Other")
nrow(data_orphans)

#> NULL