Map nest and general disturbance observations interactively.

map_dist(
  dist,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "%d/%m/%Y %H:%M",
  cluster = FALSE
)

Arguments

dist
Turtle nest and general disturbance and predation, e.g. wastd_data$nest_dist. This data contains both dist/pred recorded against nests and general signs of disturbance or predation.

wastd_url
The base URL for WAStD, default: get_wastd_url()

fmt
The desired date format, default: "d/m/Y H:M"

cluster
If TRUE, cluster map markers. Default: FALSE. Note: In some places, the aerial background layer does not provide imagery at sufficient zoom levels, and therefore restricts the map zoom at levels where the cluster markers don't expand. Switch to "Place names" to let cluster markers expand.

Value

A leaflet map

Details

Creates a Leaflet map with an interactive legend offering to toggle each disturbance class separately. The maps auto-zooms to the extent of data given.

See also

Examples

if (FALSE) {
data("wastd_data")
wastd_data$nest_dist %>% map_dist()
}