Map tracks directly from ODK data

Map tracks directly from ODK data

map_tracks_odkc(
  tracks,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "%d/%m/%Y %H:%M",
  tz = "Australia/Perth",
  cluster = FALSE,
  ts = FALSE
)

Arguments

tracks

The output of parse_turtle_nest_encounters and add_nest_labels.

sites

An sf object of sites with site_name and polygon geom, e.g. turtleviewer::turtledata$sites.

wastd_url

The base URL for WAStD, default: get_wastd_url()

fmt

The desired date format, default: "d/m/Y H:M"

tz

The lubridate timezone, default: "Australia/Perth

cluster

If TRUE, cluster map markers. Default: FALSE. Note: In some places, the aerial background layer does not provide imagery at sufficient zoom levels, and therefore restricts the map zoom at levels where the cluster markers don't expand. Switch to "Place names" to let cluster markers expand.

ts

Whether to render the data as additional timeseries (warning: slow)

Value

A leaflet map

Details

Creates a Leaflet map with an interactive legend offering to toggle each species separately. The maps auto-zooms to the extent of data given.

This map function uses data from ODK Central / ruODK using form "Turtle Track or Nest 1.0".

See also

Examples

if (FALSE) {
data("odkc_data")
map_tracks_odkc(odkc_data$tracks, sites = odkc_data$sites, cluster = TRUE)
map_tracks_odkc(odkc_data$tracks, sites = odkc_data$sites, cluster = FALSE)
map_tracks_odkc(odkc_data$tracks, sites = NULL, cluster = TRUE)
map_tracks_odkc(odkc_data$tracks, sites = NULL, cluster = FALSE)
}