Parse a wastd_api_response of area to tbl_df

Parse a wastd_api_response of area to tbl_df

Parse a wastd_api_response of animal-encounters to tbl_df

parse_surveys(wastd_api_response, wastd_url = wastdr::get_wastd_url())

parse_surveys(wastd_api_response, wastd_url = wastdr::get_wastd_url())

Arguments

wastd_api_response
A wastd_api_response of animal-encounters, e.g. get_wastd("animal-encounters")

wastd_url
The base URL for WAStD, default: get_wastd_url()

Value

A tbl_df with columns:

area_name <chr>

area_type <chr>

area_id <int>

northern_extent <num>

length_surveyed_m <int>

length_survey_roundtrip_m <int>

A tbl_df with columns:

site_name <chr>

site_type <chr>

site_id <int>

reporter <chr>

reporter_username <chr>

reporter_id <chr>

start_time <dttm> The actual start time of the survey.

end_time <dttm> The actual, automatically guessed (if 6h later), or human guessed end time.

start_comments <chr> Comments at start plus QA messages from username guessing

end_comments <chr> Comments at end plus QA messages from username guessing. A mismatch of original usernames can indicate an incorrectly picked Site Visit End point.

turtle_date <dttm> The "turtle date" of the survey.

season_week <int> The number of completed weeks since fiscal year start

iso_week <int> The number of completed weeks since calendar year start

season <int> The "turtle season" of the survey, rolling over with Fiscal Year.

source <chr> Where this record was born.

source_id <chr> The ODK record UID of the Site Visit Start.

end_source_id <chr> The ODK record UID of the Site Visit End.

status <chr> The QA status of the Survey. "Proofread" or "Curated" indicate human QA edits.

device_id <chr> The unique ID of the device the Site Visit Start was captured on.

end_device_id <chr> The unique ID of the device the Site Visit End was captured on.

status <chr> QA status of the survey.

id <int> WAStD ID of the survey.

is_production <bool> Whether this survey is a real (production) or a training survey.

absolute_admin_url <chr> The absolute URL path to edit the survey in WAStD. Append this to the base `WASTD_URL`.

start_photo_url <chr> The URL of the start photo, if given. Requires DBCA SSO when viewed.

end_photo_url <chr> The URL of the end photo, if given. Requires DBCA SSO when viewed.