Parse a wastd_api_response of taxon-conservationlisting to tbl_df

Parse a wastd_api_response of taxon-conservationlisting to tbl_df

parse_taxon_conservationlisting(
  wastd_api_response,
  wastd_url = wastdr::get_wastd_url()
)

Arguments

wastd_api_response
A wastd_api_response of taxon-conservationlisting, e.g. get_wastd("taxon-conservationlisting").

wastd_url
The base URL for WAStD, default: get_wastd_url()

Value

A tbl_df with columns:

id <int> The TSC ID of the Conservation Listing

taxon <int> The WACensus NameID of the taxon

source <int> The TSC ID for the Source of the Conservation Listing.

source_id <chr> The original ID of the source as chr. Most sources will be an integer, uuid, or any other unique ID.

scope <int> The TSC ID for the legal scope of this Conservation Listing: State of Western Australia (0), Commonwealth of Austalia (1), International (2), or Action Plan (3).

status <int> The TSC ID of the approval status, e.g. listed (80), de-listed (90), rejected (100), or in stages of approval (0..70).

proposed_on <dttm> The date on which this Conservation Listing was proposed on (GMT+08).

effective_from <dttm> The date on which this Conservation Listing was effective from (GMT+08).

effective_to <dttm> The date on which this Conservation Listing was effective_to (GMT+08).

last_reviewed_on <dttm> The date on which this Conservation Listing was last reviewed on (GMT+08).

review_due <dttm> The date on which this Conservation Listing will be due for reviewed (GMT+08).

comments <chr> Any comments. Worth a read!

category_cache <chr> A human-readable representation of the conservation category (or categories).

criteria_cache <chr> A human-readable representation of the conservation criteria.

label <chr> An auto-generated label for the Conservation Listing.

category <lst> A list if TSC conservation category IDs.

criteria <lst> A list if TSC conservation criterion IDs.