Return the number of surveys for a given site_id and season

survey_ground_covered(surveys, sid, kms, seas)

Arguments

surveys

(tibble) The output of parse_surveys.

sid

The ID of a site, e.g. 22

kms

The numbers of kilometers covered in a survey, e.g. 1.6

seas

The season as integer, e.g. 2018

See also

Examples

data("wastd_data")
one_season <- unique(wastd_data$surveys$season)[1]
one_site_id <- unique(wastd_data$surveys$site_id)[1]
some_kms <- 10
survey_ground_covered(wastd_data$surveys, one_site_id, some_kms, one_season)
#> [1] 430