Return GeoJSON features from a WAStD API endpoint

Call the WAStD API serializer's list view with given GET parameters, parse the response's features into a nested list. This function requires the WAStD API to return the results in a key features (if GeoJSON) or data (if JSON).

wastd_GET(
  serializer,
  query = list(),
  format = "json",
  max_records = NULL,
  chunk_size = 1000,
  api_url = get_wastdr_api_url(),
  api_token = get_wastdr_api_token(),
  api_un = get_wastdr_api_un(),
  api_pw = get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)

Arguments

serializer
(character) WAStD API serializer name (required) Possible values see https://tsc.dbca.wa.gov.au/api/1/

query
A list of GET parameters, default: list(). The format is specified in a separate top-level param.

format
The desired API output format, default: "json".

max_records
The maximum number of records retrieved. If left at default (NULL), all records are returned. Default: NULL.

chunk_size
The number of records to retrieve in each paginated response. A specified but smaller limit will override chunk_size. Adjust chunk_size down if getting timeouts from the API. Default: 1000.

api_url
(character) The WAStD API URL, default get_wastdr_api_url, see wastdr_setup

api_token
(character) The WAStD API token, default get_wastdr_api_token, see wastdr_setup

api_un
(character) A WAStD API username, default get_wastdr_api_un, see wastdr_setup

api_pw
(character) A WAStD API password, default get_wastdr_api_pw, see wastdr_setup

verbose
Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

An S3 object of class 'wastd_api_response' containing:

features: The sent GeoJSON features

serializer: The called serializer, e.g. 'supra'

response: The API HTTP response with all metadata

Details

See also

Examples

if (FALSE) {
track_records <- wastd_GET("turtle-nest-encounters")
tag_records <- wastd_GET("animal-encounters")
hatched_nest_records <- wastd_GET("turtle-nest-encounters",
  query = list(nest_type = "hatched-nest")
)
}