Download all turtle data from WAStD

download_wastd_turtledata(max_records = NULL, verbose = get_wastdr_verbose())

Arguments

max_records
(int) The max number of records to retrieve, default: NULL (download all).

verbose
Whether to show debug messages. Default: wastdr::get_wastdr_verbose()

Value

An S3 class "wastd_data" with items:

downloaded_on A UTC POSIXct timestamp of the data snapshot generated by Sys.time.

sites An sf object of known WAStD sites.

areas An sf object of known WAStD localities.

surveys A tibble of surveys, reconstructed by WAStD from Site Visit Start and End records (where given) and (TurtleNest)Encounters.

animals A tibble of AnimalEncounters, which includes live sightings, rescues, strandings, and other incidents, and turtle tagging. Data is included from 2016 and onwards.

tracks Turtle tracks and nests.

nest_dist Turtle Nest and General disturbance / predation.

nest_tags Turtle Nest Tags.

nest_excavations Turtle Nest Excavations - hatching and emergence success.

hatchling_morph Hatchling morphometrics from next exacavations.

nest_fans Hatchling emergence tracks - main fan.

nest_fan_outliers Hatchling emergence tracks - individual outliers.

nest_lightsources - Light sources present during hatchling emergence.

nest_loggers - Temperature loggers inside the nest. LoggerEncounters were migrated to become LoggerObservations.

linetx LineTransects of Turtle Track Tally counts.

track_tally Tally of track type/species, many per linetx.

disturbance_tally Tally of dist/pred, many per linetx.

loggers LoggerEncounters, deprecated, will be migrated to LoggerObs associated with Encounters.

Details

See also ??wastdr::export_wastd_turtledata for a detailed explanation of the data tables.

See also